Version 0.16 - move grammar source out of core
