+++ src/core/messages/index.ts
@@ -1,19 +1,18 @@
+// Grammar
+export * from "../../grammar";
+
 // Directories
 export * from "./methods";
 
 // Files
 export * from "./body";
 export * from "./digest-authentication";
-export * from "./grammar";
 export * from "./incoming-message";
 export * from "./incoming-request-message";
 export * from "./incoming-request";
 export * from "./incoming-response-message";
 export * from "./incoming-response";
-export * from "./name-addr-header";
 export * from "./outgoing-request-message";
 export * from "./outgoing-request";
 export * from "./outgoing-response";
-export * from "./parameters";
 export * from "./parser";
-export * from "./uri";
